requirements.txt
setuptools
numpy==1.26.0
chatterbox-tts==0.1.2
openai
aiogram-dialog
aiogram
jsonpickle
schedule